added feature for speakers: add/delete lections, etc

+++ commands/lection.go
@@ -30,7 +30,7 @@ const (
 	TEMPLATE_ADD_LECTION_DESCIRPTION_ERROR_NOT_YOUR = "Це не твоя лекція!"
 	TEMPLATE_LECTION_ERROR_WRONG_ID                 = "Невірно вибрана лекція"
 	TEMPLATE_WRONG_USER_ID                          = "Невідомий користувач"
-	TEMPLATE_LECTION_LIST_ITEM                      = "Лекція %d: %s\nЛектор: %s\n\n"
+	TEMPLATE_LECTION_LIST_ITEM                      = "Лекція %d: %s\nЛектор: @%s,  %s\n\n"
 	TEMPLATE_LECTION_LIST_EMPTY                     = "Поки лекцій немає"
 	TEMPLATE_DELETE_LECTION_COMPLETE                = "Лекцію успішно видалено"
 )
@@ -44,6 +44,7 @@ func nextDay(hour int) time.Time {
 }
 
 type addLection struct {
+	u           *store.User
 	step        int
 	name        string
 	description string
@@ -51,7 +52,8 @@ type addLection struct {
 }
 
 func (c *addLection) IsAllow(u *store.User) bool {
-	return u.Role == store.USER_ROLE_ADMIN
+	c.u = u
+	return (u.Role == store.USER_ROLE_ADMIN || u.Role == store.USER_ROLE_LECTOR)
 }
 
 func (c *addLection) NextStep(answer string) (*ReplyMarkup, error) {
@@ -60,6 +62,13 @@ func (c *addLection) NextStep(answer string) (*ReplyMarkup, error) {
 	}
 	switch c.step {
 	case 0:
+		// if the user is a lector = add him as an lection owner and skip the next step
+		if c.u.Role == store.USER_ROLE_LECTOR {
+			c.userID = c.u.ID
+			replyMarkup.Text = TEMPLATE_CREATE_EVENT_STEP_LECTION_NAME
+			c.step++
+			break
+		}
 		users, err := store.Users([]store.UserRole{store.USER_ROLE_ADMIN, store.USER_ROLE_LECTOR})
 		if err != nil {
 			return nil, err
@@ -129,11 +138,15 @@ func (c *addDescriptionLection) NextStep(answer string) (*ReplyMarkup, error) {
 	}
 	switch c.step {
 	case 0:
-		lections, err := store.LectionsWithoutDescriptions(c.u.ID)
+		lections, err := store.Lections(true)
 		if err != nil {
 			return nil, err
 		}
 		for _, l := range lections {
+			if (c.u.Role != store.USER_ROLE_ADMIN && c.u.ID != l.Lector.ID) || l.Description != "" {
+				// skip lections which doesn't belong to user (if he isn't admin) or it has description
+				continue
+			}
 			lText := fmt.Sprintf(TEMPLATE_LECTION_NAME, l.ID, l.Name)
 			replyMarkup.Buttons = append(replyMarkup.Buttons, lText)
 		}
@@ -153,7 +166,7 @@ func (c *addDescriptionLection) NextStep(answer string) (*ReplyMarkup, error) {
 		if err != nil {
 			return nil, err
 		}
-		if l.Lector.Username != c.u.Username {
+		if c.u.Role != store.USER_ROLE_ADMIN && l.Lector.Username != c.u.Username {
 			replyMarkup.Text = TEMPLATE_ADD_LECTION_DESCIRPTION_ERROR_NOT_YOUR
 			return replyMarkup, nil
 		}
@@ -180,8 +193,8 @@ func (c *addDescriptionLection) IsEnd() bool {
 }
 
 type lectionsList struct {
-	u     *store.User
-	empty bool
+	u                  *store.User
+	withoutDescription bool
 }
 
 func (c *lectionsList) IsEnd() bool {
@@ -190,14 +203,14 @@ func (c *lectionsList) IsEnd() bool {
 
 func (c *lectionsList) IsAllow(u *store.User) bool {
 	c.u = u
-	return u.Role == store.USER_ROLE_ADMIN
+	return (u.Role == store.USER_ROLE_ADMIN || u.Role == store.USER_ROLE_LECTOR)
 }
 
 func (c *lectionsList) NextStep(answer string) (*ReplyMarkup, error) {
 	replyMarkup := &ReplyMarkup{
 		Buttons: StandardMarkup(c.u.Role),
 	}
-	list, err := store.Lections(c.empty)
+	list, err := store.Lections(true)
 	if err != nil {
 		return nil, err
 	}
@@ -206,7 +219,15 @@ func (c *lectionsList) NextStep(answer string) (*ReplyMarkup, error) {
 		return replyMarkup, nil
 	}
 	for _, l := range list {
-		replyMarkup.Text += fmt.Sprintf(TEMPLATE_LECTION_LIST_ITEM, l.ID, l.Name, l.Lector.Name)
+		if c.withoutDescription && l.Description != "" {
+			// if we want to see only lections without descriptions and the current lection does have description - skip it
+			continue
+		}
+		if c.u.Role != store.USER_ROLE_ADMIN && c.u.ID != l.Lector.ID {
+			// skip lections which doesn't belong to user (if he isn't admin)
+			continue
+		}
+		replyMarkup.Text += fmt.Sprintf(TEMPLATE_LECTION_LIST_ITEM, l.ID, l.Name, l.Lector.Username, l.Lector.Name)
 	}
 	return replyMarkup, nil
 }
@@ -223,7 +244,7 @@ func (d *deleteLection) IsEnd() bool {
 
 func (d *deleteLection) IsAllow(u *store.User) bool {
 	d.u = u
-	return u.Role == store.USER_ROLE_ADMIN
+	return u.Role == store.USER_ROLE_ADMIN || u.Role == store.USER_ROLE_LECTOR
 }
 
 func (d *deleteLection) NextStep(answer string) (*ReplyMarkup, error) {
@@ -237,6 +258,10 @@ func (d *deleteLection) NextStep(answer string) (*ReplyMarkup, error) {
 			return nil, err
 		}
 		for _, l := range lections {
+			if d.u.Role != store.USER_ROLE_ADMIN && d.u.ID != l.Lector.ID {
+				// skip lections which doesn't belong to user (if he isn't admin)
+				continue
+			}
 			lText := fmt.Sprintf(TEMPLATE_LECTION_NAME, l.ID, l.Name)
 			replyMarkup.Buttons = append(replyMarkup.Buttons, lText)
 		}
@@ -251,6 +276,14 @@ func (d *deleteLection) NextStep(answer string) (*ReplyMarkup, error) {
 		lID, err := strconv.Atoi(matches[1])
 		if err != nil {
 			return nil, ErrWrongLection
+		}
+		l, err := store.LoadLection(lID)
+		if err != nil {
+			return nil, err
+		}
+		if d.u.Role != store.USER_ROLE_ADMIN && d.u.ID != l.Lector.ID {
+			replyMarkup.Text = TEMPLATE_ADD_LECTION_DESCIRPTION_ERROR_NOT_YOUR
+			return replyMarkup, nil
 		}
 		err = store.DeleteLection(lID)
 		if err != nil {
